fix(moviePage): destructure + ui fix
