+++ client/src/features/reviews/components/ReviewElement.tsx
@@ -5,6 +5,7 @@ import { Box, Typography, Paper, Button, Stack, Rating } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
 import ReplyIcon from '@mui/icons-material/Reply';
+import { formatDate } from '@/shared/utils/dataFormatter';
 
 type Props = {
   review: Review;
@@ -36,19 +37,29 @@ export default function ReviewElement({
         borderRadius: 5,
       }}
     >
-      <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-        <Typography fontSize={20}>
-          <Box component="span" fontWeight="bold">
-            {isUser ? 'You' : review.reviewer.username}
-          </Box>{' '}
-          rated
+      <Box display={'flex'} justifyContent="space-between">
+        <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+          <Typography fontSize={20}>
+            <Box component="span" fontWeight="bold">
+              {isUser ? 'You' : review.reviewer.username}
+            </Box>{' '}
+            rated
+          </Typography>
+          <Rating
+            name="read-only-rating"
+            value={review.stars}
+            precision={0.5}
+            readOnly
+          />
+        </Box>
+        <Typography>
+          {formatDate(review.createdAt)}{' '}
+          {review.createdAt !== review.updatedAt && (
+            <Typography component="span" fontWeight="bold">
+              (edited)
+            </Typography>
+          )}
         </Typography>
-        <Rating
-          name="read-only-rating"
-          value={review.stars}
-          precision={0.5}
-          readOnly
-        />
       </Box>
 
       <Typography mt={1}>{review.content}</Typography>
